Animals

Named vector of animals

Details

animals is a named character vector of animals, with each element a character string of variable length specifying an ASCII animal. Note that some have unicode characters that won't play well on some operating systems.

Examples

animals
#>  [1] "alligator"    "ant"          "anxiouscat"   "bat"          "bat2"        
#>  [6] "beavis"       "behindcat"    "bigcat"       "blowfish"     "buffalo"     
#> [11] "cat"          "chicken"      "chuck"        "clippy"       "cow_borg"    
#> [16] "cow_dead"     "cow_greedy"   "cow_sleepy"   "cow_tired"    "cow_wired"   
#> [21] "cow_young"    "cow"          "daemon"       "dragon"       "duck"        
#> [26] "duckling"     "egret"        "endlesshorse" "facecat"      "fish"        
#> [31] "frog"         "ghost"        "goldfish"     "grumpycat"    "hypnotoad"   
#> [36] "longcat"      "longtailcat"  "monkey"       "mushroom"     "owl"         
#> [41] "pig"          "poop"         "pumpkin"      "rabbit"       "shark"       
#> [46] "shortcat"     "signbunny"    "smallcat"     "snowman"      "spider"      
#> [51] "squirrel"     "squirrel2"    "stegosaurus"  "stretchycat"  "trilobite"   
#> [56] "turkey"       "whale"        "wolf"         "yoda"        
cat(animal_fetch('cow'))
#> 
#>         ^__^ 
#>         (oo)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
cat(animal_fetch('chicken'))
#>          _
#>        _/ }
#>       `>' \
#>       `|   \
#>        |   /'-.     .-.
#>         \'     ';`--' .'
#>          \'.    `'-./
#>           '.`-..-;`
#>             `;-..'
#>             _| _|
#>             /` /` [nosig]
cat(animal_fetch('chuck'))
#> 
#> 
#>                       @@@@@@@@@@@@@@
#>                    @@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>  @             @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>   @@           @@@@@@@@@@@@@@@@@@     @@@@@@@@@@@@@@@@@@@@@
#>     @@@@@@@@@@@@@@@@    @@ @@@@         @@  @@@@@@@@@@@@@@@
#>        @@@@@@@@@@@@@      @@@@          @@@@@@@@@@@@@@@@@@@
#>           @@@@@@ @@@@@    @@@   @       @@@@@@@@@@ @@@@@@@
#>               @@     @ @                @@ @@@@@@@@   @@@
#>                                               @@@@@@@  @
#>                 @     @@                   @ @@@@@@@@@
#>                              @               @@@@@@@@ @@
#>                      @@@@@@@@@@          @  @@@@@
#>                   @@@@@@@@@@ @@           @@@@@  @
#>                  @@@@        @@ @          @@@  @
#>                   @@@           @          @@@
#>                    @@@@@@@   @@ @          @@@@@
#>                     @@@@@@@@@@            @@@@
#>                      @@@@@@ @     @     @@@@@@@@
#>                   @@ @ @@@@@@@@@ @@@@@@@@@@@@@@@@@
#>                 @@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@
#>       @@@@@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#> @@    @@@@  @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ [nosig]
cat(animal_fetch('clippy'))
#> 
#>     _
#>    / \
#>    | |
#>    @ @
#>   || ||
#>   || ||
#>   |\_/|
#>   \___/ GB
cat(animal_fetch('poop'))
#> 
#>      (   )
#>   (   ) (
#>    ) _   )
#>     ( \_
#>   _(_\ \)__
#>  (____\ ___)) [nosig]
cat(animal_fetch('bigcat'))
#> 
#>                 \`*-.
#>                  )  _`-.
#>                 .  : `. .
#>                 : _   '  \
#>                 ; *` _.   `*-._
#>                 `-.-'          `-.
#>                   ;       `       `.
#>                   :.       .       \
#>                   .\  .   :   .-'   .
#>                   '  `+.;  ;  '      :
#>                   :  '  |    ;       ;-.
#>                   ; '   : :`-:     _.`* ;
#>                .*' /  .*' ; .*`- +'  `*'
#>      [bug]     `*-*   `*-*  `*-*'