Sling messages and warnings with flair

Sling messages and warnings with flair

say(
  what = "Hello world!",
  by = "cat",
  type = NULL,
  what_color = NULL,
  by_color = NULL,
  length = 18,
  fortune = NULL,
  ...
)

Arguments

what

(character) What do you want to say? See Details.

by

(character) Type of thing, one of cow, chicken, chuck, clippy, poop, bigcat, ant, pumpkin, ghost, spider, rabbit, pig, snowman, frog, hypnotoad, shortcat, longcat, fish, signbunny, facecat, behindcat, stretchycat, anxiouscat, longtailcat, cat, trilobite, shark, buffalo, grumpycat, smallcat, yoda, mushroom, endlesshorse, bat, bat2, turkey, monkey, daemon, egret, duckling, duck, owl, squirrel, squirrel2, goldfish, alligator, stegosaurus, whale, wolf, or rms for Richard Stallman. Alternatively, use "random" to have your message spoken by a random character. We use match.arg() internally, so you can use unique parts of words that don't conflict with others, like "g" for "ghost" because there's no other animal that starts with "g".

type

(character) One of message (default), warning, print (default in non-interactive mode), or string (returns string). If multiple colors are supplied to what_color or by_color, type cannot be warning. (This is a limitation of the multicolor packcage :/.) If run in non-interactive mode default type is print, so that output goes to stdout rather than stderr, where messages and warnings go.

what_color

(character or crayon function) One or more crayon-suported text color(s) or crayon style function to color what. You might try colors() or ?rgb for ideas. Use "rainbow" for c("red", "orange", "yellow", "green", "blue", "purple").

by_color

(character or crayon function) One or more crayon-suported text color(s) or crayon style function to color who. Use "rainbow" for c("red", "orange", "yellow", "green", "blue", "purple")

length

(integer) Length of longcat. Ignored if other animals used.

fortune

An integer specifying the row number of fortunes.data. Alternatively which can be a character and grep is used to try to find a suitable row.

...

Further args passed on to fortunes::fortune()

what

You can put in any phrase you like to the what parameter, OR you can type in one of a few special phrases that do particular things. They are:

"catfact": A random cat fact from https://catfact.ninja

"fortune": A random quote from an R coder, from fortunes library

"time": Print the current time

"rms": Prints a random 'fact' about Richard Stallman from the rmsfact::rmsfact() package. Best paired with by = "rms".

The fillerama API was an option ("arresteddevelopment", "doctorwho", "dexter", "futurama", "holygrail", "simpsons", "starwars", "loremipsum"), a text generator with quotes from shows and movies. However, the API is currently down so you cannot use this option anymore.

by

Note that if you choose by='hypnotoad' the quote is forced to be, as you could imagine, 'All Glory to the HYPNO TOAD!'. For reference see http://knowyourmeme.com/memes/hypnotoad

signbunny: It's not for sure known who invented signbunny, but this article http://www.vox.com/2014/9/18/6331753/sign-bunny-meme-explained thinks they found the first use in this tweet: https://twitter.com/wei_bluebear/status/329101645780770817

trilobite: from http://www.retrojunkie.com/asciiart/animals/dinos.htm (site down though)

Note to Windows users: there are some animals (shortcat, longcat, fish, signbunny, stretchycat, anxiouscat, longtailcat, grumpycat, mushroom) that are not available because they use non-ASCII characters that don't display properly in R on Windows.

Examples

say()
#> 
#>  -------------- 
#> Hello world! 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say("what")
#> 
#>  -------------- 
#> what 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say("meow", "cat", what_color = "blue")
#> 
#>  -------------- 
#> meow 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say('time')
#> 
#>  -------------- 
#> 2022-09-23 18:55:12 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say('time', "poop", by_color = "cyan", what_color = "pink")
#> 
#>  ----- 
#> 2022-09-23 18:55:12 
#>  ------ 
#>     \   
#>      \
#>      (   )
#>   (   ) (
#>    ) _   )
#>     ( \_
#>   _(_\ \)__
#>  (____\ ___)) [nosig]

library(jsonlite)
say("hello world",
    by = "hypnotoad",
    what_color = c("royalblue1", "tomato2"),
    by_color = c("rainbow", "rainbow"))
#>  ----- 
#> hello world 
#>  ------
#>     \          ,'``.._   ,'``.
#>      \        :,--._:)\,:,._,.:
#>       \       :`--,''   :`...';\
#>                `,'       `---'  `.
#>                /                 :
#>               /                   \
#>             ,'                     :\.___,-.
#>            `...,---'``````-..._    |:       \
#>              (                 )   ;:    )   \  _,-.
#>               `.              (   //          `'    \
#>                :               `.//  )      )     , ;
#>              ,-|`.            _,'/       )    ) ,' ,'
#>             (  :`.`-..____..=:.-':     .     _,' ,'
#>              `,'\ ``--....-)='    `._,  \  ,') _ '``._
#>           _.-/ _ `.       (_)      /     )' ; / \ \`-.'
#>          `--(   `-:`.     `' ___..'  _,-'   |/   `.)
#>              `-. `.`.``-----``--,  .'
#>                |/`.\`'        ,','); SSt
#>                    `         (/  (/
#>   
say("whats up",
    what_color = crayon::cyan$bgMagenta,
    by_color = c("salmon1", "springgreen"))
#>  -------------- 
#> whats up 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   

say("who you callin chicken", "chicken")
#> 
#>  ----- 
#> who you callin chicken 
#>  ------ 
#>     \   
#>      \
#>          _
#>        _/ }
#>       `>' \
#>       `|   \
#>        |   /'-.     .-.
#>         \'     ';`--' .'
#>          \'.    `'-./
#>           '.`-..-;`
#>             `;-..'
#>             _| _|
#>             /` /` [nosig]
#>   
say("ain't that some shit", "poop")
#> 
#>  ----- 
#> ain't that some shit 
#>  ------ 
#>     \   
#>      \
#>      (   )
#>   (   ) (
#>    ) _   )
#>     ( \_
#>   _(_\ \)__
#>  (____\ ___)) [nosig]
say("icanhazpdf?", "cat")
#> 
#>  -------------- 
#> icanhazpdf? 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say("boo!", "pumpkin")
#> 
#>  ----- 
#> boo! 
#>  ------ 
#>     \   
#>      \
#>                   ___
#>                ___)__|_
#>           .-*'          '*-,
#>          /      /|   |\     \
#>         ;      /_|   |_\     ;
#>         ;   |\           /|  ;
#>         ;   | ''--...--'' |  ;
#>          \  ''---.....--''  /
#>           ''*-.,_______,.-*'  [nosig]
#>   
say("hot diggity", "frog")
#> 
#>  ----- 
#> hot diggity 
#>  ------ 
#>     \   
#>      \
#>         (.)_(.)
#>      _ (   _   ) _
#>     / \/`-----'\/ \
#>   __\ ( (     ) ) /__
#>   )   /\ \._./ /\   (
#>    )_/ /|\   /|\ \_(  [nosig]
#>   
say("fortune", "spider")
#> 
#>  ----- 
#> Henrik Bengtsson: Is there a way to turn off the (annoying) beep that occurs when one calls the locator() command and clicks the mouse? [...]
#> Brian D. Ripley: It's a feature of the windows() device. Windows is fond of beeping, and I just mute the sound.
#>  Henrik Bengtsson and Brian D. Ripley
#>  R-help
#>  July 2001 
#>  ------ 
#>     \   
#>      \
#>               |
#>               |
#>               |
#>              __
#>           | /  \ |
#>          \_\\  //_/
#>           .'/()\'.
#>            \\  //  [nosig]
#>   
say("fortune", "facecat")
#> 
#>  -------------- 
#> 'They' did write documentation that told you that Perl was needed, but 'they' can't read it for you.
#>  Brian D. Ripley
#>  R-help
#>  February 2001 
#>  --------------
#>     \
#>       \
#>          /\ /\
#>          (O o)
#>         =(:^:)=
#>            U      [nosig]
#>   
say("fortune", "behindcat")
#> 
#>  -------------- 
#> You must realize that R is written by experts in statistics and statistical computing who, despite popular opinion, do not believe that everything in SAS and SPSS is worth copying. Some things done in such packages, which trace their roots back to the days of punched cards and magnetic tape when fitting a single linear model may take several days because your first 5 attempts failed due to syntax errors in the JCL or the SAS code, still reflect the approach of "give me every possible statistic that could be calculated from this model, whether or not it makes sense". The approach taken in R is different. The underlying assumption is that the useR is thinking about the analysis while doing it.
#>  Douglas Bates
#>  in reply to the suggestion to include type III sums of squares and lsmeans in base R to make it more similar to SAS or SPSS
#>  R-help
#>  March 2007 
#>  --------------
#>       \
#>         \
#>           \
#>             |\___/|
#>             )     (
#>            =\     /=
#>              )===(
#>             /     \
#>             |     |
#>            /       \
#>            \       /
#>       jgs   \__  _/
#>               ( (
#>                ) )
#>               (_(
#>   
say("fortune", "smallcat")
#> 
#>  -------------- 
#> Bret Collier: I hope this is not a uniformed question, but I am a little lost.
#> Peter Dalgaard: Don't worry, they all look alike... ;-)
#>  Bret Collier and Peter Dalgaard
#>  R-help
#>  April 2004 
#>  --------------
#>     \
#>      \
#>       \
#>          /\_/\
#>         ( o.o )
#>          > ^ <      [nosig]
#> 
say("fortune", "monkey")
#> 
#>  ------------- 
#> On average, any data manipulation that can be described in a sentence or two of English can be programmed in one line in R. If you find yourself writing a long 'for' loop to do something that sounds simple, take a step back and research if an existing combination of functions can easily handle your request.
#>  Erik Iverson
#>  R-help
#>  June 2010 
#>  -------------- 
#>               \   
#>                \  
#>                 \
#> 
#>                   .="=.
#>                 _/.-.-.\_     _
#>                ( ( o o ) )    ))
#>                 |/  "  \|    //
#>                  \'---'/    //
#>            jgs   /`"""`\\  ((
#>                 / /_,_\ \\  \\
#>                 \_\_'__/  \  ))
#>                 /`  /`~\   |//
#>                /   /    \  /
#>           ,--`,--'\/\    /
#>           '-- "--'  '--'
say("fortune", "egret")
#> 
#>  ----- 
#> I recently read the small print on the academic license our site has for SAS. You have to:
#> 
#>   1 inform SAS of any taught courses that use SAS,
#>   2 inform SAS of any research projects using SAS,
#>   3 allow SAS to refer to your institution as a SAS user,
#>   4 allow SAS to review your taught courses,
#>   5 ensure your courses are taught using qualified personnel,
#>   6 give SAS your first-born male offspring.
#> 
#> I spoke to our site's licensing supremos and they say they've never heard of anyone complying with 1 or 2. Point 4 sounds like petty fiddling in our educational business, and point 5 left 'qualified' undefined. Point 6 doesn't bother me since I don't have kids.
#>  Barry Rowlingson
#>  R-help
#>  February 2008 
#>  ------ 
#>     \   
#>      \  
#>       \
#>        \   _,
#>       -==<' `
#>           ) /
#>          / (_.
#>         |  ,-,`\
#>          \\   \ \
#>           `\,  \ \
#>            ||\  \`|,
#>  jgs      _|| `=`-'
#>          ~~`~`
say("rms", "rms")
#> 
#> 
#>  ----- 
#> Some people wear Linus Torvalds pyjamas to bed, Linus Torvalds wears Richard Stallman pyjamas. 
#>  ------ 
#>     \   
#>      \
#>                     @@@@@@ @
#>                   @@@@     @@
#>                  @@@@ =   =  @@ 
#>                 @@@ @ _   _   @@ 
#>                  @@@ @(0)|(0)  @@ 
#>                 @@@@   ~ | ~   @@
#>                 @@@ @  (o1o)    @@
#>                @@@    #######    @
#>                @@@   ##{+++}##   @@
#>               @@@@@ ## ##### ## @@@@
#>               @@@@@#############@@@@
#>              @@@@@@@###########@@@@@@
#>             @@@@@@@#############@@@@@
#>             @@@@@@@### ## ### ###@@@@
#>              @ @  @              @  @
#>                @                    @

# Vary type of output, default calls message()
say("hell no!")
#> 
#>  -------------- 
#> hell no! 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say("hell no!", type="warning")
#> Warning: 
#>  -------------- 
#> hell no! 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say("hell no!", type="string")
#> [1] "\n -------------- \nhell no! \n --------------\n    \\\n      \\\n        \\\n            |\\___/|\n          ==) ^Y^ (==\n            \\  ^  /\n             )=*=(\n            /     \\\n            |     |\n           /| | | |\\\n           \\| | |_|/\\\n      jgs  //_// ___/\n               \\_)\n  "

# Using fortunes
say(what="fortune")
#> 
#>  -------------- 
#> Any doubts about R's big-league status should be put to rest, now that we have a Sudoku Puzzle Solver. Take that, SAS!
#>  David Brahm
#>  announcing the sudoku package
#>  R-packages
#>  January 2006 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
## you don't have to pass anything to the `what` parameter if `fortune` is
## not null
say(fortune=10)
#> 
#>  -------------- 
#> Overall, SAS is about 11 years behind R and S-Plus in statistical capabilities (last year it was about 10 years behind) in my estimation.
#>  Frank Harrell (SAS User, 1969-1991)
#>  R-help
#>  September 2003 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say(fortune=100)
#> 
#>  -------------- 
#> I'm not sure I'd trust any computer recommendation from 1976, no matter how famous the authors are.
#>  Peter Dalgaard
#>  after Samuel Edward Kemp cited a recommendation about nonlinear least squares computer programs from 'Box-Jenkins, 1976'
#>  R-help
#>  January 2005 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say(fortune='whatever')
#> 
#>  -------------- 
#> Justin: Is there a function that just does whatever I'm thinking (aka whatever my homework question is...)?
#> Joshua Ulrich: That's the magic_pony function.
#>  Justin and Joshua Ulrich
#>  stackoverflow.com
#>  June 2013 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say(fortune=7)
#> 
#>  -------------- 
#> What we have is nice, but we need something very different.
#>  Robert Gentleman
#>  Statistical Computing 2003, Reisensburg
#>  June 2003 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
say(fortune=45)
#> 
#>  -------------- 
#> It seems to me against the spirit of Open Source software to attempt to monitor distribution. We could ask R to 'call home' on first use (in the way e.g. pine does) but I suspect many users would find that objectionable.
#>  Brian D. Ripley
#>  in a discussion about trying to estimate the number of R users
#>  R-help
#>  April 2004 
#>  --------------
#>     \
#>       \
#>         \
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   

# Using catfacts
# say("catfact", "cat")

# The hypnotoad
say(by="hypnotoad")
#> 
#>  ----- 
#> All Glory to the HYPNO TOAD! 
#>  ------
#>     \          ,'``.._   ,'``.
#>      \        :,--._:)\,:,._,.:
#>       \       :`--,''   :`...';\
#>                `,'       `---'  `.
#>                /                 :
#>               /                   \
#>             ,'                     :\.___,-.
#>            `...,---'``````-..._    |:       \
#>              (                 )   ;:    )   \  _,-.
#>               `.              (   //          `'    \
#>                :               `.//  )      )     , ;
#>              ,-|`.            _,'/       )    ) ,' ,'
#>             (  :`.`-..____..=:.-':     .     _,' ,'
#>              `,'\ ``--....-)='    `._,  \  ,') _ '``._
#>           _.-/ _ `.       (_)      /     )' ; / \ \`-.'
#>          `--(   `-:`.     `' ___..'  _,-'   |/   `.)
#>              `-. `.`.``-----``--,  .'
#>                |/`.\`'        ,','); SSt
#>                    `         (/  (/
#>   

# Trilobite
say(by='trilobite')
#> 
#>  -------------- 
#> Hello world! 
#>  --------------
#>     \
#>       \
#>         \
#>           _____
#>        .'` ,-. `'.
#>       /   ([ ])   \
#>      /.-""`(`)`""-.\
#>       <'```(.)```'>
#>       <'```(.)```'>
#>        <'``(.)``'>
#>    sk   <``\_/``>
#>          `'---'`
#>   

# Shark
say('Q: What do you call a solitary shark\nA: A lone shark', by='shark')
#> 
#>  -------------- 
#> Q: What do you call a solitary shark
#> A: A lone shark 
#>  --------------
#>     \
#>       \
#>         \
#>               /""-._
#>               .       '-,
#>                :          '',
#>                 ;      *     '.
#>                  ' *         () '.
#>                    \               \
#>                     \      _.---.._ '.
#>                     :  .' _.--''-''  \ ,'
#>         .._           '/.'             . ;
#>         ; `-.          ,                \'
#>          ;   `,         ;              ._\
#>           ;    \     _,-'                ''--._
#>           :    \_,-'                          '-._
#>           \ ,-'                       .          '-._
#>           .'         __.-'';            \...,__       '.
#>         .'      _,-'        \              \   ''--.,__  '\
#>         /    _,--' ;         \              ;           \^.}
#>         ;_,-' )     \  )\      )            ;
#>              /       \/  \_.,-'             ;
#>             /                              ;
#>          ,-'  _,-'''-.    ,-.,            ;      PFA
#>       ,-' _.-'        \  /    |/'-._...--'
#>      :--``             )/
#>   '
#>   

# Buffalo
say('Q: What do you call a single buffalo?\nA: A buffalonely', by='buffalo')
#> 
#>  -------------- 
#> Q: What do you call a single buffalo?
#> A: A buffalonely 
#>  --------------
#>     \
#>       \
#>         \
#>                    _.-````'-,_
#>          _,.,_ ,-'`           `'-.,_
#>        /)     (                   '``-.
#>       ((      ) )                      `\
#>         \)    (_/                        )\
#>         |       /)           '    ,'    / \
#>         `\    ^'            '     (    /  ))
#>           |      _/\ ,     /    ,,`\   (  "`
#>           \Y,   |   \  \  | ````| / \_ \
#>             `)_/      \  \  )    ( >  ( >
#>                        \( \(     |/   |/
#>           mic & dwb  /_(/_(    /_(  /_(
#>     

# Clippy
say(fortune=59, by="clippy")
#> 
#> 
#>  ----- 
#> Let's not kid ourselves: the most widely used piece of software for statistics is Excel.
#>  Brian D. Ripley
#>  'Statistical Methods Need Software: A View of Statistical Computing'
#>  Opening lecture RSS 2002, Plymouth
#>  September 2002 
#>  ------ 
#>     \   
#>      \   __
#>    / \
#>    | |
#>    @ @
#>   || ||
#>   || ||
#>   |\_/|
#>   \___/ GB