Animals

Named vector of animals

Usage

animals

Format

An object of class character of length 56.

Details

animals is a named character vector of animals, with each element a character string of variable length specifying an ASCII animal. Note that some have unicode characters that won't play well on some operating systems.

Examples

cat(animals['cow'])
#> 
#>         ^__^ 
#>         (oo)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
cat(animals['chicken'])
#>          _
#>        _/ }
#>       `>' \
#>       `|   \
#>        |   /'-.     .-.
#>         \'     ';`--' .'
#>          \'.    `'-./
#>           '.`-..-;`
#>             `;-..'
#>             _| _|
#>             /` /` [nosig]
#>  
cat(animals[['chuck']])
#> 
#> 
#>                       @@@@@@@@@@@@@@
#>                    @@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>  @             @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>   @@           @@@@@@@@@@@@@@@@@@     @@@@@@@@@@@@@@@@@@@@@
#>     @@@@@@@@@@@@@@@@    @@ @@@@         @@  @@@@@@@@@@@@@@@
#>        @@@@@@@@@@@@@      @@@@          @@@@@@@@@@@@@@@@@@@
#>           @@@@@@ @@@@@    @@@   @       @@@@@@@@@@ @@@@@@@
#>               @@     @ @                @@ @@@@@@@@   @@@
#>                                               @@@@@@@  @
#>                 @     @@                   @ @@@@@@@@@
#>                              @               @@@@@@@@ @@
#>                      @@@@@@@@@@          @  @@@@@
#>                   @@@@@@@@@@ @@           @@@@@  @
#>                  @@@@        @@ @          @@@  @
#>                   @@@           @          @@@
#>                    @@@@@@@   @@ @          @@@@@
#>                     @@@@@@@@@@            @@@@
#>                      @@@@@@ @     @     @@@@@@@@
#>                   @@ @ @@@@@@@@@ @@@@@@@@@@@@@@@@@
#>                 @@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@
#>       @@@@@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#> @@    @@@@  @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ [nosig]
#>   
cat(animals[['clippy']])
#> 
#>     _
#>    / \
#>    | |
#>    @ @
#>   || ||
#>   || ||
#>   |\_/|
#>   \___/ GB
cat(animals[['poop']])
#> 
#>      (   )
#>   (   ) (
#>    ) _   )
#>     ( \_
#>   _(_\ \)__
#>  (____\ ___)) [nosig]
cat(animals[['bigcat']])
#> 
#>                 \`*-.
#>                  )  _`-.
#>                 .  : `. .
#>                 : _   '  \
#>                 ; *` _.   `*-._
#>                 `-.-'          `-.
#>                   ;       `       `.
#>                   :.       .       \
#>                   .\  .   :   .-'   .
#>                   '  `+.;  ;  '      :
#>                   :  '  |    ;       ;-.
#>                   ; '   : :`-:     _.`* ;
#>                .*' /  .*' ; .*`- +'  `*'
#>      [bug]     `*-*   `*-*  `*-*'
#>     
for (animal in animals) cat(animal, sep = "\n")
#> 
#>         ^__^ 
#>         (oo)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (==)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         ($$)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (**)\ ________ 
#>         (__)\         )\ /\ 
#>          U   ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (xx)\ ________ 
#>         (__)\         )\ /\ 
#>          U   ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (--)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (OO)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#> 
#>         ^__^ 
#>         (..)\ ________ 
#>         (__)\         )\ /\ 
#>              ||------w|
#>              ||      ||
#>          _
#>        _/ }
#>       `>' \
#>       `|   \
#>        |   /'-.     .-.
#>         \'     ';`--' .'
#>          \'.    `'-./
#>           '.`-..-;`
#>             `;-..'
#>             _| _|
#>             /` /` [nosig]
#>  
#> 
#> 
#>                       @@@@@@@@@@@@@@
#>                    @@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                   @@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                  @@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                 @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>                @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>  @             @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#>   @@           @@@@@@@@@@@@@@@@@@     @@@@@@@@@@@@@@@@@@@@@
#>     @@@@@@@@@@@@@@@@    @@ @@@@         @@  @@@@@@@@@@@@@@@
#>        @@@@@@@@@@@@@      @@@@          @@@@@@@@@@@@@@@@@@@
#>           @@@@@@ @@@@@    @@@   @       @@@@@@@@@@ @@@@@@@
#>               @@     @ @                @@ @@@@@@@@   @@@
#>                                               @@@@@@@  @
#>                 @     @@                   @ @@@@@@@@@
#>                              @               @@@@@@@@ @@
#>                      @@@@@@@@@@          @  @@@@@
#>                   @@@@@@@@@@ @@           @@@@@  @
#>                  @@@@        @@ @          @@@  @
#>                   @@@           @          @@@
#>                    @@@@@@@   @@ @          @@@@@
#>                     @@@@@@@@@@            @@@@
#>                      @@@@@@ @     @     @@@@@@@@
#>                   @@ @ @@@@@@@@@ @@@@@@@@@@@@@@@@@
#>                 @@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@
#>       @@@@@@  @@@@@   @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
#> @@    @@@@  @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ [nosig]
#>   
#> 
#>     _
#>    / \
#>    | |
#>    @ @
#>   || ||
#>   || ||
#>   |\_/|
#>   \___/ GB
#> 
#> 
#>      (   )
#>   (   ) (
#>    ) _   )
#>     ( \_
#>   _(_\ \)__
#>  (____\ ___)) [nosig]
#> 
#>                 \`*-.
#>                  )  _`-.
#>                 .  : `. .
#>                 : _   '  \
#>                 ; *` _.   `*-._
#>                 `-.-'          `-.
#>                   ;       `       `.
#>                   :.       .       \
#>                   .\  .   :   .-'   .
#>                   '  `+.;  ;  '      :
#>                   :  '  |    ;       ;-.
#>                   ; '   : :`-:     _.`* ;
#>                .*' /  .*' ; .*`- +'  `*'
#>      [bug]     `*-*   `*-*  `*-*'
#>     
#> 
#>        '\__
#>       (o )     ___
#>       <>(_)(_)(___)
#>         < < > >
#>         ' ' ` `  [nosig]
#>   
#> 
#>                   ___
#>                ___)__|_
#>           .-*'          '*-,
#>          /      /|   |\     \
#>         ;      /_|   |_\     ;
#>         ;   |\           /|  ;
#>         ;   | ''--...--'' |  ;
#>          \  ''---.....--''  /
#>           ''*-.,_______,.-*'  [nosig]
#>   
#> 
#>      .-.
#>     (o o)
#>     | O \
#>      \   \
#>       `~~~' [nosig]
#>   
#> 
#>               |
#>               |
#>               |
#>              __
#>           | /  \ |
#>          \_\\  //_/
#>           .'/()\'.
#>            \\  //  [nosig]
#>   
#> 
#>       ( )_( )
#>       (='.'=)
#>       (^)_(^) [nosig]
#>   
#> 
#>        _//| .-~~~-.
#>      _/oo  }       }-@
#>     ('')_  }       |
#>      `--'| { }--{  }
#>           //_/  /_/ [nosig]
#>   
#> 
#>      _[_]_
#>       (")
#>   >--( : )--<
#>     (__:__) [nosig]
#>   
#> 
#>         (.)_(.)
#>      _ (   _   ) _
#>     / \/`-----'\/ \
#>   __\ ( (     ) ) /__
#>   )   /\ \._./ /\   (
#>    )_/ /|\   /|\ \_(  [nosig]
#>   
#> 
#>                 ,'``.._   ,'``.
#>                 :,--._:)\,:,._,.:
#>                 :`--,''   :`...';\
#>                `,'       `---'  `.
#>                /                 :
#>               /                   \
#>             ,'                     :\.___,-.
#>            `...,---'``````-..._    |:       \
#>              (                 )   ;:    )   \  _,-.
#>               `.              (   //          `'    \
#>                :               `.//  )      )     , ;
#>              ,-|`.            _,'/       )    ) ,' ,'
#>             (  :`.`-..____..=:.-':     .     _,' ,'
#>              `,'\ ``--....-)='    `._,  \  ,') _ '``._
#>           _.-/ _ `.       (_)      /     )' ; / \ \`-.'
#>          `--(   `-:`.     `' ___..'  _,-'   |/   `.)
#>              `-. `.`.``-----``--,  .'
#>                |/`.\`'        ,','); SSt
#>                    `         (/  (/
#>   
#> 
#>     .ﾊ,,ﾊ
#>     ( ﾟωﾟ)
#>     |つ  つ
#>     U " U
#>         [BoingBoing]
#>     
#> 
#>     .ﾊ,,ﾊ
#>     ( ﾟωﾟ)
#>     |つ  つ
#> %s
#>     U "  U
#>         [BoingBoing]
#>     
#> 
#>   ><((((º>  ><((((º>  ><((((º>  ><((((º>  ><((((º>
#>       Kiyoko Gotanda
#>     
#> 
#> (\__/) ||
#> (•ㅅ•) ||
#> /   づ
#>           [nosig]
#>     
#> 
#>          /\ /\
#>          (O o)
#>         =(:^:)=
#>            U      [nosig]
#>   
#> 
#>             |\___/|
#>             )     (
#>            =\     /=
#>              )===(
#>             /     \
#>             |     |
#>            /       \
#>            \       /
#>       jgs   \__  _/
#>               ( (
#>                ) )
#>               (_(
#>   
#> 
#>                         ,/|         _.--‛‛^``-...___.._.,;
#>                       /, \‛.     _-‛          ,--,,,--‛‛‛
#>                      {  \    `_-‛‛       ‛    /}‛
#> Jill                    `;;‛             ;   ; ;
#>                   ._.--‛‛     ._,,, _..‛  .;.‛
#>                   (,_....----‛‛‛     (,..--‛‛
#>   
#> 
#>         /\_/\         _
#>        /``   \       / )
#>        |n n   |__   ( (
#>       =(Y =.‛`   `\  \ \
#>       {`"`        \  ) )
#>       {       /    |/ /
#>        \\   ,(     / /
#>         ) ) /-‛\  ,_.‛
#>   jgs  (,(,/ ((,,/
#>   
#> 
#>      /\-/\
#>     /a a  \                                 _
#>    =\ Y  =/-~~~~~~-,_______________________/ )
#>      ‛^--‛          ________________________/
#>        \           /
#>        ||  |---‛\  \
#>   jgs  (_(__|   ((__|
#>   
#> 
#>             |\___/|
#>           ==) ^Y^ (==
#>             \  ^  /
#>              )=*=(
#>             /     \
#>             |     |
#>            /| | | |\
#>            \| | |_|/\
#>       jgs  //_// ___/
#>                \_)
#>   
#> 
#>           _____
#>        .'` ,-. `'.
#>       /   ([ ])   \
#>      /.-""`(`)`""-.\
#>       <'```(.)```'>
#>       <'```(.)```'>
#>        <'``(.)``'>
#>    sk   <``\_/``>
#>          `'---'`
#>   
#> 
#>               /""-._
#>               .       '-,
#>                :          '',
#>                 ;      *     '.
#>                  ' *         () '.
#>                    \               \
#>                     \      _.---.._ '.
#>                     :  .' _.--''-''  \ ,'
#>         .._           '/.'             . ;
#>         ; `-.          ,                \'
#>          ;   `,         ;              ._\
#>           ;    \     _,-'                ''--._
#>           :    \_,-'                          '-._
#>           \ ,-'                       .          '-._
#>           .'         __.-'';            \...,__       '.
#>         .'      _,-'        \              \   ''--.,__  '\
#>         /    _,--' ;         \              ;           \^.}
#>         ;_,-' )     \  )\      )            ;
#>              /       \/  \_.,-'             ;
#>             /                              ;
#>          ,-'  _,-'''-.    ,-.,            ;      PFA
#>       ,-' _.-'        \  /    |/'-._...--'
#>      :--``             )/
#>   '
#>   
#> 
#>                    _.-````'-,_
#>          _,.,_ ,-'`           `'-.,_
#>        /)     (                   '``-.
#>       ((      ) )                      `\
#>         \)    (_/                        )\
#>         |       /)           '    ,'    / \
#>         `\    ^'            '     (    /  ))
#>           |      _/\ ,     /    ,,`\   (  "`
#>           \Y,   |   \  \  | ````| / \_ \
#>             `)_/      \  \  )    ( >  ( >
#>                        \( \(     |/   |/
#>           mic & dwb  /_(/_(    /_(  /_(
#>     
#> 
#>       ﾊ _ ﾊ
#>       ಠ X ಠ
#>   
#> 
#>          /\_/\
#>         ( o.o )
#>          > ^ <      [nosig]
#> 
#> 
#> 
#>                    ____
#>                 _.' :  `._
#>             .-.'`.  ;   .'`.-.
#>    __      / : ___\ ;  /___ ; \      __
#>   ,'_ ""--.:__;".-.";: :".-.":__;.--"" _`,
#>   :' `.t""--.. '<@.`;_  ',@>` ..--""j.' `;
#>        `:-.._J '-.-'L__ `-- ' L_..-;'
#>           "-.__ ;  .-"  "-.  : __.-"
#>              L ' /.------.\ ' J
#>              "-.   "--"   .-"
#>              __.l"-:_JL_;-";.__
#>          .-j/'.;  ;""""  / .'\"-.
#>          .' /:`. "-.:     .-" .';  `.
#>       .-"  / ;  "-. "-..-" .-"  :    "-.
#>   .+"-.  : :      "-.__.-"      ;-._   \
#>   ; \  `.; ;                    : : "+. ;
#>   :  ;   ; ;                    : ;  : \:
#>   ;  :   ; :                    ;:   ;  :
#>   : \  ;  :  ;                  : ;  /  ::
#>   ;  ; :   ; :                  ;   :   ;:
#>   :  :  ;  :  ;                : :  ;  : ;
#>   ;\    :   ; :                ; ;     ; ;
#>   : `."-;   :  ;              :  ;    /  ;
#>  ;    -:   ; :              ;  : .-"   :
#>   :\     \  :  ;            : \.-"      :
#>   ;`.    \  ; :            ;.'_..--  / ;
#>   :  "-.  "-:  ;          :/."      .'  :
#>    \         \ :          ;/  __        :
#>     \       .-`.\        /t-""  ":-+.   :
#>      `.  .-"    `l    __/ /`. :  ; ; \  ;
#>        \   .-" .-"-.-"  .' .'j \  /   ;/
#>         \ / .-"   /.     .'.' ;_:'    ;
#>   :-""-.`./-.'     /    `.___.'
#>                \ `t  ._  /  bug
#>                 "-.t-._:'
#>   
#> 
#>                 ________
#>            __--´      ° `--__
#>        __-´     °      °     `-__
#>      (´    °    °          °     `)
#>      (° °|    |°         ° |    | )
#>       `'''''''''`|'''''|´''''''''´
#>                  |     |
#>                  |:::::|
#>                /:|:::::|:\
#>               /::|:::::|::\
#>                  |     |
#>                  |^   ^|
#>                  |  _  | [FK]
#>                  |_____|
#> 
#> 
#>        ,
#>     _,,)\.~,,._
#>      (()`  ``)\))),,_
#>       |     \ ''((\)))),,_          ____
#>       |6`   |   ''((\())) "-.____.-"    `-.-,
#>       |    .'\    ''))))'                  \)))
#>       |   |   `.     ''                     ((((
#>       \, _)     \/                          |))))
#>        `'        |                          (((((
#>                  \                  |       ))))))
#>                   `|    |           ,\     /((((((
#>                    |   / `-.______.<  \   |  )))))
#>                    |   |  /         `. \  \  ((((
#>                    |  / \ |           `.\  | (((
#>                    \  | | |             )| |  ))
#>                     | | | |             || |  '   [endless.horse]
#>                     | | | |             || |
#> 
#>         __.--'\     \.__./     /'--.__
#>     _.-'       '.__.'    '.__.'       '-._
#>   .'                                      '.
#>  /                                          \
#> |                                            |
#> |                                            |
#>  \         .---.              .---.         /
#>   '._    .'     '.''.    .''.'     '.    _.'
#>      '-./            \  /           \.-'
#>                       ''mrf
#> 
#> 
#> _____________________                              _____________________
#> `-._                 \           |\__/|           /                 _.-'
#>     \                 \          |    |          /                 /
#>      \                 `-_______/      \_______-'                 /
#>       |                                                          |
#>       |                                                          |
#>       |                                                          |
#>       /                                                          \
#>      /_____________                                  _____________\
#>                    `----._                    _.----'
#>                           `--.            .--'
#>                               `-.      .-'
#>                                  \    / :F_P:
#>                                   \  /
#>                                    \/
#> 
#> 
#>                .--.
#>               /} p \             /}
#>              `~)-) /           /` }
#>               ( / /          /`}.' }
#>                / / .-'""-.  / ' }-'}
#>               / (.'       \/ '.'}_.}
#>              |            `}   .}._}
#>              |     .-=-';   } ' }_.}
#>              \    `.-=-;'  } '.}.-}
#>               '.   -=-'    ;,}._.}
#>                 `-,_  __.'` '-._}
#>               jgs   `|||
#>                    .=='=,
#> 
#> 
#>                   .="=.
#>                 _/.-.-.\_     _
#>                ( ( o o ) )    ))
#>                 |/  "  \|    //
#>                  \'---'/    //
#>            jgs   /`"""`\\  ((
#>                 / /_,_\ \\  \\
#>                 \_\_'__/  \  ))
#>                 /`  /`~\   |//
#>                /   /    \  /
#>           ,--`,--'\/\    /
#>           '-- "--'  '--'
#> 
#>             ,        ,
#>             /(        )`
#>             \ \___   / |
#>             /- _  `-/  '
#>            (/\/ \ \   /\
#>            / /   | `    
#>            O O   ) /    |
#>            `-^--'`<     '
#>           (_.)  _  )   /
#>            `.___/`    /
#>              `-----' /
#> <----.     __ / __   \
#> <----|====O)))==) \) /====
#> <----'    `--' `.__,' \
#>              |        |
#>               \       /
#>         ______( (_  / \______
#>       ,'  ,-----'   |        \
#>       `--{__________)        \/ [nosig]
#> 
#> 
#>            _,
#>       -==<' `
#>           ) /
#>          / (_.
#>         |  ,-,`\
#>          \\   \ \
#>           `\,  \ \
#>            ||\  \`|,
#>  jgs      _|| `=`-'
#>          ~~`~`
#> 
#> 
#>         >o  .
#>          ===    [ab]
#> 
#> 
#>          __
#>         /o \
#>       <=   |         ==
#>         |__|        /===
#>         |   \______/  =
#>         \     ====   /
#>          \__________/     [ab]
#> 
#> 
#>        /\___/\
#>        {o}{o}|
#>        \ v  /|
#>        |    \ \
#>         \___/_/       [ab]
#>           | |
#> 
#> 
#>                       . .
#>                       |\|\_
#>                       /  ^ \
#>                      /  _-_/°
#>    \\\\\\\\\       /   / \
#>   ////////////   /  \ / ||
#>  \\\\\\\\\\\\\\ /   /\\ \\
#> ////////////////   /  \\ \\
#>  \\\\\\\\\\\\\/   /  / `` ``
#>      /////////   \  /  \
#>  ML     \\\\\\___/_/___/
#> 
#> 
#>                       . .
#>                       |\|\_
#>                       /  @ \
#>                      /  _-_/°
#>    \\\\\\\\\       /   / \
#>   ////////////   /  \ / ||
#>  \\\\\\\\\\\\\\ /   /\\ \\
#> ////////////////   /  \\ \\
#>  \\\\\\\\\\\\\/   /  / `` ``
#>      /////////   \  /  \
#>  ML     \\\\\\___/_/___/
#> 
#> 
#>             /`-._
#>            /_,.._`:-
#>       ,.-'  ,   `-:..-')
#>       : o ):';      _  {
#>        `-._ `'__,.-'\`-.)
#>            `\\  \,.-'`   Max
#> 
#> 
#>            .-._   _ _ _ _ _ _ _ _
#> .-''-.__.-'00  '-' ' ' ' ' ' ' ' '-.
#> '.___ '    .   .--_'-' '-' '-' _'-' '._
#> V: V 'vv-'   '_   '.       .'  _..' '.'.
#>   '=.____.=_.--'   :_.__.__:_   '.   : :
#>            (((____.-'        '-.  /   : :
#> snd                          (((-'\ .' /
#>                            _____..'  .'
#>                           '-._____.-'
#> 
#> 
#>                          .       .
#>                         / `.   .' \
#>                 .---.  <    > <    >  .---.
#>                 |    \  \ - ~ ~ - /  /    |
#>                  ~-..-~             ~-..-~
#>              \~~~\.'                    `./~~~/
#>    .-~~^-.    \__/                        \__/
#>  .'  O    \     /               /       \  \
#> (_____,    `._.'               |         }  \/~~~/
#>  `----.          /       }     |        /    \__/
#>        `-.      |       /      |       /      `. ,~~|
#>            ~-.__|      /_ - ~ ^|      /- _      `..-'   f: f:
#>                 |     /        |     /     ~-.     `-. _||_||_
#>                 |_____|        |_____|         ~ - . _ _ _ _ _>
#> 
#> 
#>      .-'
#> '--./ /     _.---.
#> '-,  (__..-`       \
#>    \          .     |
#>     `,.__.   ,__.--/
#>      '._/_.'___.-`
#> 
#> 
#>                     .
#>                    / V\
#>                   / `  /
#>                  <<   |
#>                  /    |
#>                /      |
#>              /        |
#>            /    \  \ /
#>           (      ) | |
#>   ________|   _/_  | |
#> <__________\______)\__)
#> 
#>          __------~~-,
#>       ,'            ,
#>      /               \
#>     /                :
#>    |                  '
#>    |                  |
#>    |                  |
#>     |   _--           |
#>     _| =-.     .-.   ||
#>     o|/o/       _.   |
#>     /  ~          \ |
#>   (____\@)  ___~    |
#>      |_===~~~.`    |
#>   _______.--~     |
#>   \________       |
#>            \      |
#>          __/-___-- -__
#>         /            _ \
#> 
#> 
#>             |    .
#>         .   |L  /|
#>     _ . |\ _| \--+._/| .
#>    / ||\| Y J  )   / |/| ./
#>   J  |)'( |        ` F`.'/
#> -<|  F         __     .-<
#>   | /       .-'. `.  /-. L___
#>   J \      <    \  | | O\|.-'
#> _J \  .-    \/ O | | \  |F
#> '-F  -<_.     \   .-'  `-' L__
#> __J  _   _.     >-'  )._.   |-'
#> `-|.'   /_.           \_|   F
#>   /.-   .                _.<
#>  /'    /.'             .'  `\
#>   /L  /'   |/      _.-'-\
#>  /'J       ___.---'\|
#>    |\  .--' V  | `. `
#>    |/`. `-.     `._)
#>       / .-.\
#>       \ (  `\
#>        `.\
#>