Sling messages and warnings with flair

If you are familiar with cowsay on the cli, then you know what this is, but for R. If not, read below. Why? Why not?

Details

Contributors:

Amanda Dobbyn

Andreas Brandmaier

Andy Teucher

Carson Sievert

Franz-Sebastian Krah

Guangchuang Yu

Karl Broman

Kiyoko Gotanda

Lucy D'Agostino McGowan

Noam Ross

Paolo Sonego

Philipp Boersch-Supan

Rich FitzJohn

Scott Chamberlain

Thomas Leeper

Tyler Rinker

Marion Louveaux

There is only one function in the cowsay package. Simplicity for the win.

Say something:

say('hello world!')