Endless horse

Each time you press enter, the horse keeps going...and going...

endless_horse(
  what = "Hello world!",
  endless = TRUE,
  wait = 0.5,
  what_color = NULL,
  horse_color = NULL
)

Arguments

what

(character) What do you want to say? See details.

endless

(logical) Should horse be enless, you better say yes. Default: TRUE

wait

How long to wait between leg segments (time grows geometrically after the first iteration in order to keep the horse on screen for a while, but it will keep going forever. Or until you hit escape/Ctrl-C depending on your platform).

what_color

(character or crayon function) A crayon-suported text color or crayon style function to color what. You might try colors() or ?rgb for ideas.

horse_color

(character or crayon function) A crayon-suported text color or crayon style function to color your steed.

Examples

if (FALSE) {
endless_horse()
endless_horse(endless = FALSE)
}